DarrenS: Refactoring of DAOs into subfolders based on type
