+++ src/main/java/edu/lwtech/csd297/teachersfirst/Security.java
@@ -7,7 +7,7 @@ import javax.servlet.http.HttpServletRequest;
 
 import org.apache.logging.log4j.*;
 
-import edu.lwtech.csd297.teachersfirst.daos.MemberSqlDAO;
+import edu.lwtech.csd297.teachersfirst.daos.sql.MemberSqlDAO;
 import edu.lwtech.csd297.teachersfirst.obj.*;
 
 public class Security {
